# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "description": "An API for managing users in a database",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Get all users",
        "description": "Get all users from the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved all users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve users from the database"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Insert new user",
          "description": "Insert a new user into the database",
          "parameters": [
            {
              "name": "firstName",
              "in": "formData",
              "required": true,
              "type": "string"
            },
            {
              "name": "lastName",
              "in": "formData",
              "required": true,
              "type": "string"
            },
            {
              "name": "email",
              "in": "formData",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "201": {
              "description": "Successfully inserted user into the database"
            },
            "500": {
              "description": "Failed to insert user into the database"
            }
          }
        }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a user by his id",
        "description": "Get a user by their ID from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Failed to find user with the specified ID"
          },
          "500": {
            "description": "Failed to retrieve user from the database"
          }
        }
      }
    },
    "/users/{id}/bachelors/": {

      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get bachelors for user and paginate them",
        "description": "Get bachelors for the user with id = ID from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved bachelors",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Failed to find bachelors with the specified ID"
          },
          "500": {
            "description": "Failed to retrieve bachelors from the database"
          }
        }
      }
    },
    "/users/{id}/filteredBachelors": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Get bachelors for user and paginate them",
        "description": "Get bachelors for the user with id = ID from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "min_distance",
            "in": "formData",
            "required": true,
            "type": "integer",
            "minimum": 0
          },
          {
            "name": "max_distance",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "min_age",
            "in": "formData",
            "required": true,
            "type": "integer",
            "minimum": 18
          },
          {
            "name": "max_age",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "min_fame",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "min_common_interests",
            "in": "formData",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved bachelors",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "404": {
            "description": "Failed to find bachelors with the specified ID"
          },
          "500": {
            "description": "Failed to retrieve bachelors from the database"
          }
        }
      }
    },
    "/users/{id}/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user without sensitive data",
        "description": "Get a user by their ID from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Failed to find user with the specified ID"
          },
          "500": {
            "description": "Failed to retrieve user from the database"
          }
        }
      }
    },
    "/users/{id}/liked": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a list of liked users by logged in user",
        "description": "Get a user by their ID from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Failed to find user with the specified ID"
          },
          "500": {
            "description": "Failed to retrieve user from the database"
          }
        }
      }
    },
    "/users/{id}/matched": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a list of matched users by logged in user",
        "description": "Get a user by their ID from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Failed to find user with the specified ID"
          },
          "500": {
            "description": "Failed to retrieve user from the database"
          }
        }
      }
    },
    "/users/{id}/update": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update anything in a user",
        "description": "Update user from the database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "int"
          },
          {
            "name": "first_name",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "last_name",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "age",
            "in": "formData",
            "type": "integer"
          },
          {
            "name": "sex",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "sex_orientation",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "city",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "country",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "interests",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "bio",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "active",
            "in": "formData",
            "type": "boolean"
          },
          {
            "name": "fame_rating",
            "in": "formData",
            "type": "integer"
          },
          {
            "name": "report_count",
            "in": "formData",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Failed to find user with the specified ID"
          },
          "500": {
            "description": "Failed to update user from the database"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "age": {
          "type": "integer"
        },
        "sex": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "birthdate": {
          "type": "string"
        },
        "interests": {
          "type": "string"
        },
        "photos": {
          "type": "string"
        },
        "bio": {
          "type": "string"
        }
      }
    },
...
  }
}